Fix C++ warning: remove spurious ulong cast.

+++ src/basemath/FlxqE.c
@@ -448,7 +448,7 @@ FlxqE_find_order(GEN f, GEN h, GEN bound, GEN B, GEN a4, GEN T, ulong p)
     if (r)
     {
       while (tx[r] == k && r) r--;
-      for (r++; tx[r] == k && r <= (ulong)s; r++)
+      for (r++; tx[r] == k && r <= s; r++)
       {
         long j = ti[r]-1;
         GEN Q = FlxqE_add(FlxqE_mul(F, stoi(j), a4, T, p), fh, a4, T, p);
